inject: stop using LD_PRELOAD; use patchelf exclusively
LD_PRELOAD is problematic on Android due to AT_SECURE; to work around
AT_SECURE we'd need to patch a lot of SELinux policies.

Using patchelf is at least simpler than working through all those
SELinux policies. The only downside is that newer versions of patchelf
kind of break on Android, but the 0.9 version works just fine.
