+++ native/inject/jni/Android.mk
@@ -1,19 +1,6 @@
 LOCAL_PATH := $(call my-dir)
 
 include $(CLEAR_VARS)
-LOCAL_MODULE := hooked_entry
-
-LOCAL_C_INCLUDES := \
-    $(LOCAL_PATH)/common/include
-
-LOCAL_SRC_FILES := \
-    common/entry.cpp
-
-LOCAL_LDLIBS := -llog
-    
-include $(BUILD_EXECUTABLE)
-
-include $(CLEAR_VARS)
 LOCAL_MODULE := libgatekeeperd_hook
 
 LOCAL_C_INCLUDES := \
@@ -25,6 +12,19 @@ LOCAL_SRC_FILES := \
 LOCAL_LDLIBS := -llog
     
 include $(BUILD_SHARED_LIBRARY)
+
+include $(CLEAR_VARS)
+LOCAL_MODULE := app_process_wrapper
+
+LOCAL_C_INCLUDES := \
+    $(LOCAL_PATH)/common/include
+
+LOCAL_SRC_FILES := \
+    app_process_hook/wrapper.cpp
+
+LOCAL_LDLIBS := -llog
+
+include $(BUILD_EXECUTABLE)
 
 include $(CLEAR_VARS)
 LOCAL_MODULE := libapp_process_hook
